Encapsulate "inflight-op" checking into BatchIteratorBackpressureExecutor
This should fix a race condition which could cause a UPDATE or DELETE
(by query) to get stuck.

Before a `pendingItemsCount` `AtomicInteger` was used to track if there
is any operation still active. This `pendingItemsCount` was shared
accross components, one component increased it, another decreased it. It
seems like there was a race condition in that the count could get
decremented too late. This caused the operation to get stuck as the
`BatchIteratorBackpressureExecutor` would wait for another operation to
complete which never happened.